PDF Scripts/requirements.txt
pdfplumber
python-docx
Pillow
pytesseract 